Prevent the login process from updating the DOM.
Though this is a neat trick, in the event a user backs out and doesn't continue
with the Persona login process, the DOM is left in a funny state that doesn't,
in fact, reflect reality. The user is no longer logging in.

Commenting out this code let's the user attempt another login, should they
choose to, without having to reload the page to "repair" the DOM.